added delete app script
