+++ bin/push_app.py
@@ -11,51 +11,9 @@ import json
 import os
 from argparse import ArgumentParser
 import logging
-import requests
-from requests.auth import HTTPBasicAuth
+from common import make_request, get_endpoint
 
 logging.basicConfig(level=logging.INFO)
-
-
-def make_request(url, access_id, access_key, files=None, is_first_submission=False, **kwargsr):
-    """
-    Utility function to return make app push request
-    :param url:
-    :param access_id:
-    :param access_key:
-    :param files:
-    :param is_first_submission:
-    :param kwargsr:
-    :return: status,data where  status = boolean indicating success or failure of the call, data is an object depending on the API call
-    """
-    status = None
-    logging.debug("Calling %s: " % url)
-    if is_first_submission:
-        response = requests.post(url=url, auth=HTTPBasicAuth(
-            access_id, access_key), files=files)
-
-    else:
-        response = requests.get(url=url, auth=HTTPBasicAuth(access_id, access_key))
-    status = True if response.status_code == 200 else False
-
-    logging.debug("Got response code: %s Content:%s" % (response.status_code, response.content))
-    try:
-        if len(response.content) > 0:
-            try:
-                response = response.content.strip("\" '")
-            except TypeError:
-                # python 3
-                response = response.content.strip(b"\" '")
-            data = json.loads(response)
-        else:
-            data = {'error': "true", "message": response.status_code}
-    except Exception as e:
-        logging.error(e)
-        data = {'error': "true", "message": response}
-    if 'error' in data:
-        # old API call so need to check for error again
-        status = (data['error'] == "false")
-    return status, data
 
 
 def push_app_api(api_url, sourcefile, manifestfile, access_id, access_key):
@@ -63,7 +21,7 @@ def push_app_api(api_url, sourcefile, manifestfile, access_id, access_key):
     with open(sourcefile, 'rb') as sf:
         with open(manifestfile, 'rb') as mf:
             status, response = make_request(api_url, access_id, access_key, files={'appDef': sf, 'manifest': mf},
-                                            is_first_submission=True)
+                                            method="POST")
             return status, response['message']
 
 
@@ -72,14 +30,14 @@ def push_app_api_v2(import_app_url, sourcefile, manifestfile, access_id, access_
         with open(manifestfile, 'rb') as mf:
             status, response = make_request(import_app_url, access_id, access_key,
                                             files={'appDefinition': sf, 'appManifest': mf},
-                                            is_first_submission=True)
+                                            method="POST")
             if status:
                 logging.info("Got app push job id: %s" % response['id'])
                 job_id = response["id"]
                 job_status_url = "%s/%s/status" % (import_app_url, job_id)
                 waiting = True
                 while waiting:
-                    status, response = make_request(job_status_url, access_id, access_key)
+                    status, response = make_request(job_status_url, access_id, access_key, method="GET")
                     logging.info(response)
                     if response['status'] == "InProgress":
                         waiting = True
@@ -122,25 +80,6 @@ def is_newformat(filename):
                 return True
             else:
                 return False
-
-
-def get_endpoint(deployment):
-    deploy_map = {
-        "syd": "au",
-        "mum": "in",
-        "dub": "eu",
-        "fra": "de",
-        "fed": "fed",
-        "jp": "jp",
-        "mon": "ca",
-        "us2": "us2"
-    }
-    if deployment in ("us1", "prod"):
-        return "https://api.sumologic.com/api/v1/content/app"
-    if deployment in deploy_map:
-        return "https://api.%s.sumologic.com/api/v1/content/app" % deploy_map[deployment]
-    else:
-        return 'https://%s-api.sumologic.net/api/v1/content/app' % deployment
 
 
 def get_final_api_endpoint(start_endpoint, is_foldersync_format):
